# internal/handler/subscription/get_all.go
package subscription

import (
	"net/http"

	"github.com/RSODA/subscribe-manager/internal/dto"
	"github.com/gin-gonic/gin"
)

// @Summary      Список подписок
// @Tags         subscriptions
// @Produce      json
// @Success      200  {array}   dto.SubscriptionResponse
// @Failure      500  {object}  dto.InternalErrorResponse
// @Router       /subscriptions [get]
func (h *Handler) GetAll(c *gin.Context) {
	res, err := h.service.GetAll(c.Request.Context())
	if err != nil {
		h.writeError(c, err)
		return
	}

	c.JSON(http.StatusOK, dto.NewSubscriptionResponses(res))
}